feat: MODE 전환을 route 방법으로 변경 #18
