+++ src/layouts/Main/Main.tsx
@@ -1,4 +1,3 @@
-import { useEffect } from "react";
 import { Outlet } from "react-router-dom";
 import styled from "styled-components";
 import Footer from "../Footer/Footer";
@@ -6,10 +5,6 @@ import Header from "../Header/Header";
 import Section from "../Section/Section";
 
 const Main = () => {
-  useEffect(() => {
-    console.log('side effect main');
-
-  }, []);
   return (
     <StyledMainContainer>
       <Header />
