+++ src/pages/Calendar/CalendarPage.tsx
@@ -1,0 +1,9 @@
+const CalendarPage = () => {
+  return (
+    <div>
+      Cal
+    </div>
+  );
+};
+
+export default CalendarPage;
