+++ src/pages/Memo/MemoPage.tsx
@@ -1,0 +1,9 @@
+const MemoPage = () => {
+  return (
+    <div>
+      Memo
+    </div>
+  );
+};
+
+export default MemoPage;
