+++ src/router.tsx
@@ -5,11 +5,15 @@ import AuthRouter from '@/routes/Auth';
 import NotFoundPage from '@/pages/NotFound/NotFoundPage';
 import LoadingSpinner from '@/components/LoadingSpinner/LoadingSpinner';
 import { useLogin } from '@/utils/hooks';
-import Main from './layouts/Main/Main';
+import Main from '@/layouts/Main/Main';
 
 const HomePage = lazy(() => import('@/pages/Home/HomePage'));
 const LandingPage = lazy(() => import('@/pages/Landing/LandingPage'));
 const ProfilePage = lazy(() => import('@/pages/Profile/ProfilePage'));
+const CalendarPage = lazy(() => import('@/pages/Calendar/CalendarPage'));
+const MemoPage = lazy(() => import('@/pages/Memo/MemoPage'));
+const TodoPage = lazy(() => import('@/pages/Todo/TodoPage'));
+const CustomPage = lazy(() => import('@/pages/Custom/CustomPage'));
 const TermsPage = lazy(() => import('@/pages/Terms/TermsPage'));
 
 const AppRouter = () => {
@@ -33,6 +37,38 @@ const AppRouter = () => {
               element={
                 <AuthRouter>
                   <ProfilePage />
+                </AuthRouter>
+              }
+            />
+            <Route
+              path="/calendar"
+              element={
+                <AuthRouter>
+                  <CalendarPage />
+                </AuthRouter>
+              }
+            />
+            <Route
+              path="/memo"
+              element={
+                <AuthRouter>
+                  <MemoPage />
+                </AuthRouter>
+              }
+            />
+            <Route
+              path="/todo"
+              element={
+                <AuthRouter>
+                  <TodoPage />
+                </AuthRouter>
+              }
+            />
+            <Route
+              path="/custom"
+              element={
+                <AuthRouter>
+                  <CustomPage />
                 </AuthRouter>
               }
             />
